+++ ArkMonitor/Misc/ArkColors.swift
@@ -11,13 +11,14 @@ import UIKit
 public struct ArkColors {
     
     // 9E9E9E
-    static public let gray = UIColor(red: 158/255, green: 158/255, blue: 158/255, alpha: 1.0)
+    static public let gray = UIColor(red: 132/255, green: 132/255, blue: 132/255, alpha: 1.0)
     
     // F1F8E9
     static public let lightGreen = UIColor(red: 241/255, green: 248/255, blue: 233/255, alpha: 1.0)
     
-    // 1E88E5
-    static public let blue = UIColor(red: 30/255, green: 136/255, blue: 229/255, alpha: 1.0)
+    // 1e73be
+    static public let blue = UIColor(red: 30/255, green: 115/255, blue: 190/255, alpha: 1.0)
     
-    
+    // 4c4c4c
+    static public let darkGray = UIColor(red: 76/255, green: 76/255, blue: 76/255, alpha: 1.0)
 }
